+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/jobs/SpiderJob.java
@@ -19,6 +19,7 @@
  */
 package org.zaproxy.addon.automation.jobs;
 
+import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -27,13 +28,21 @@ import java.util.concurrent.TimeUnit;
 import org.apache.commons.httpclient.URI;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.control.Control;
+import org.parosproxy.paros.extension.history.ExtensionHistory;
+import org.parosproxy.paros.model.HistoryReference;
+import org.parosproxy.paros.model.Model;
+import org.parosproxy.paros.network.ConnectionParam;
+import org.parosproxy.paros.network.HttpMessage;
+import org.parosproxy.paros.network.HttpSender;
+import org.parosproxy.paros.network.HttpStatusCode;
 import org.zaproxy.addon.automation.AutomationEnvironment;
 import org.zaproxy.addon.automation.AutomationJob;
 import org.zaproxy.addon.automation.AutomationProgress;
+import org.zaproxy.addon.automation.ContextWrapper;
 import org.zaproxy.zap.extension.spider.ExtensionSpider;
 import org.zaproxy.zap.extension.spider.SpiderScan;
-import org.zaproxy.zap.model.Context;
 import org.zaproxy.zap.model.Target;
+import org.zaproxy.zap.utils.ThreadUtils;
 
 public class SpiderJob extends AutomationJob {
 
@@ -56,6 +65,8 @@ public class SpiderJob extends AutomationJob {
 
     private String contextName;
     private String url;
+
+    private UrlRequester urlRequester = new UrlRequester(this.getName());
 
     public SpiderJob() {}
 
@@ -106,44 +117,50 @@ public class SpiderJob extends AutomationJob {
     public void runJob(
             AutomationEnvironment env, LinkedHashMap<?, ?> jobData, AutomationProgress progress) {
 
-        Context context;
+        ContextWrapper context;
         if (contextName != null) {
-            context = env.getContext(contextName);
+            context = env.getContextWrapper(contextName);
             if (context == null) {
                 progress.error(
                         Constant.messages.getString(
-                                "automation.error.context.unknown",
-                                env.getUrlStringForContext(context)));
+                                "automation.error.context.unknown", contextName));
                 return;
             }
         } else {
-            context = env.getDefaultContext();
+            context = env.getDefaultContextWrapper();
         }
         URI uri = null;
         try {
             if (url != null) {
                 uri = new URI(url, true);
-            } else {
-                uri = new URI(env.getUrlStringForContext(context).toString(), true);
             }
         } catch (Exception e1) {
-            progress.error(
-                    Constant.messages.getString(
-                            "automation.error.context.badurl",
-                            env.getUrlStringForContext(context)));
+            progress.error(Constant.messages.getString("automation.error.context.badurl", url));
             return;
         }
 
-        Target target = new Target(context);
+        // Request all specified URLs
+        for (String url : context.getUrls()) {
+            this.urlRequester.requestUrl(url, progress);
+        }
+
+        if (env.isTimeToQuit()) {
+            // Failed to access one of the URLs
+            return;
+        }
+
+        Target target = new Target(context.getContext());
         target.setRecurse(true);
         List<Object> contextSpecificObjects = new ArrayList<>();
-        contextSpecificObjects.add(uri);
+        if (uri != null) {
+            contextSpecificObjects.add(uri);
+        }
 
         int scanId = this.getExtSpider().startScan(target, null, contextSpecificObjects.toArray());
 
         long endTime = Long.MAX_VALUE;
         if (maxDuration > 0) {
-            // The spider should stop, if it doesnt we will stop it (after a few seconds leeway
+            // The spider should stop, if it doesnt we will stop it (after a few seconds leeway)
             endTime =
                     System.currentTimeMillis()
                             + TimeUnit.MINUTES.toMillis(maxDuration)
@@ -192,6 +209,15 @@ public class SpiderJob extends AutomationJob {
         }
     }
 
+    /**
+     * Only for use by unit tests
+     *
+     * @param urlRequester the UrlRequester to use
+     */
+    protected void setUrlRequester(UrlRequester urlRequester) {
+        this.urlRequester = urlRequester;
+    }
+
     @Override
     public boolean isExcludeParam(String param) {
         switch (param) {
@@ -235,5 +261,78 @@ public class SpiderJob extends AutomationJob {
     @Override
     public String getParamMethodName() {
         return OPTIONS_METHOD_NAME;
+    }
+
+    public static class UrlRequester {
+
+        private final HttpSender httpSender;
+        private final String requester;
+
+        public UrlRequester(String requester) {
+            this.requester = requester;
+            httpSender =
+                    new HttpSender(
+                            Model.getSingleton().getOptionsParam().getConnectionParam(),
+                            true,
+                            HttpSender.SPIDER_INITIATOR);
+        }
+
+        public void requestUrl(String url, AutomationProgress progress) {
+            // Request the URL
+            try {
+                final HttpMessage msg = new HttpMessage(new URI(url, true));
+                httpSender.sendAndReceive(msg, true);
+
+                if (msg.getResponseHeader().getStatusCode() != HttpStatusCode.OK) {
+                    progress.error(
+                            Constant.messages.getString(
+                                    "automation.error.spider.url.notok",
+                                    requester,
+                                    url,
+                                    msg.getResponseHeader().getStatusCode()));
+                    return;
+                }
+
+                ExtensionHistory extHistory =
+                        Control.getSingleton()
+                                .getExtensionLoader()
+                                .getExtension(ExtensionHistory.class);
+                extHistory.addHistory(msg, HistoryReference.TYPE_SPIDER);
+
+                ThreadUtils.invokeAndWait(
+                        () ->
+                                // Needs to be done on the EDT
+                                Model.getSingleton()
+                                        .getSession()
+                                        .getSiteTree()
+                                        .addPath(msg.getHistoryRef()));
+            } catch (UnknownHostException e1) {
+                ConnectionParam connectionParam =
+                        Model.getSingleton().getOptionsParam().getConnectionParam();
+                if (connectionParam.isUseProxyChain()
+                        && connectionParam.getProxyChainName().equalsIgnoreCase(e1.getMessage())) {
+                    progress.error(
+                            Constant.messages.getString(
+                                    "automation.error.spider.url.badhost.proxychain",
+                                    requester,
+                                    url,
+                                    e1.getMessage()));
+                } else {
+                    progress.error(
+                            Constant.messages.getString(
+                                    "automation.error.spider.url.badhost",
+                                    requester,
+                                    url,
+                                    e1.getMessage()));
+                }
+            } catch (Exception e1) {
+                progress.error(
+                        Constant.messages.getString(
+                                "automation.error.spider.url.failed",
+                                requester,
+                                url,
+                                e1.getMessage()));
+            }
+        }
     }
 }