+++ addOns/reports/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Maintenance changes.
+- Handle multiple context URLs in automation.
 
 ## [0.3.0] - 2021-05-06
 ### Added